Prepared crop demo for NTCrop
