+++ NTKitDemo/Base.lproj/Main.storyboard
@@ -67,10 +67,10 @@
             </objects>
             <point key="canvasLocation" x="-98" y="479"/>
         </scene>
-        <!--Tile View Controller-->
+        <!--Tile Demo View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController storyboardIdentifier="tileView" id="BYZ-38-t0r" customClass="TileViewController" customModule="NTKitDemo" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="tileView" id="BYZ-38-t0r" customClass="TileDemoViewController" customModule="NTKitDemo" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="y3c-jy-aDJ"/>
                         <viewControllerLayoutGuide type="bottom" id="wfy-db-euE"/>
@@ -108,10 +108,10 @@
             </objects>
             <point key="canvasLocation" x="558" y="479"/>
         </scene>
-        <!--Image View Controller-->
+        <!--Image Demo View Controller-->
         <scene sceneID="GYF-6k-VJc">
             <objects>
-                <viewController storyboardIdentifier="imageView" id="t51-Cg-ZbL" customClass="ImageViewController" customModule="NTKitDemo" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="imageView" id="t51-Cg-ZbL" customClass="ImageDemoViewController" customModule="NTKitDemo" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="rRM-MG-Fod"/>
                         <viewControllerLayoutGuide type="bottom" id="eGk-uP-RDn"/>
@@ -150,10 +150,10 @@
             </objects>
             <point key="canvasLocation" x="558" y="1135"/>
         </scene>
-        <!--Image View With Effects View Controller-->
+        <!--Image View With Effects Demo View Controller-->
         <scene sceneID="l6L-aw-Yy9">
             <objects>
-                <viewController storyboardIdentifier="imageViewEffects" id="05r-hr-FYl" customClass="ImageViewWithEffectsViewController" customModule="NTKitDemo" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="imageViewEffects" id="05r-hr-FYl" customClass="ImageViewWithEffectsDemoViewController" customModule="NTKitDemo" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="G5H-c1-h3p"/>
                         <viewControllerLayoutGuide type="bottom" id="cHy-SH-SLu"/>
@@ -191,6 +191,37 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="1xG-HB-Mm6" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="558" y="1811"/>
+        </scene>
+        <!--Crop Demo View Controller-->
+        <scene sceneID="iRs-WG-1bT">
+            <objects>
+                <viewController storyboardIdentifier="cropView" id="vjA-pv-HR3" customClass="CropDemoViewController" customModule="NTKitDemo" customModuleProvider="target" sceneMemberID="viewController">
+                    <layoutGuides>
+                        <viewControllerLayoutGuide type="top" id="iBq-OO-n4j"/>
+                        <viewControllerLayoutGuide type="bottom" id="mns-On-uZk"/>
+                    </layoutGuides>
+                    <view key="view" contentMode="scaleToFill" id="6Pa-L0-aP5">
+                        <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="IWm-CR-Trx">
+                                <rect key="frame" x="258" y="562" width="84" height="30"/>
+                                <state key="normal" title="Close Demo"/>
+                                <connections>
+                                    <action selector="closePressed:" destination="vjA-pv-HR3" eventType="touchUpInside" id="wvx-OD-sWy"/>
+                                </connections>
+                            </button>
+                        </subviews>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                        <constraints>
+                            <constraint firstItem="mns-On-uZk" firstAttribute="top" secondItem="IWm-CR-Trx" secondAttribute="bottom" constant="8" id="fzt-R1-aU4"/>
+                            <constraint firstItem="IWm-CR-Trx" firstAttribute="centerX" secondItem="6Pa-L0-aP5" secondAttribute="centerX" id="ymt-Y4-S2m"/>
+                        </constraints>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="a1X-fe-O5m" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="558" y="2476"/>
         </scene>
     </scenes>
 </document>
